requirements.txt
python-can==4.4.2
pytest==9.0.2
